chore(l2): separete ProverServer and ProofSender (#2478)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Currently the ProverServer component have the responsibility for both
act as a server for the ProverClient (i.e., send blocks to prove and
receive proofs) and send proofs to the L1 contract to verify blocks.
This tasks can be parallel and decoupled one from the other.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
A new struct `L1ProofSender` is created that periodically checks if
there're new proofs to send to the L1, removing that job from the
`ProverServer`. Also, components were renamed for better clarity.
Note that the config names were not changed as there's a WIP PR (#2501)
doing a full refactor of it

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>