fix(core): Made engine_forkchoiceUpdatedV3 second parameter optional (#2575)
**Motivation**

This PR makes it so that when parsing engine_forkchoiceUpdatedV3 the
second parameter isn't required. This came to light while testing
odometer #2507, which sent the updates without the second parameter.
This change makes it more conforming with the spec.

**Description**

Made it so that the second optional parameter in
engine_forkchoiceUpdatedV3 ( Payload attributes) isn't required to be
sent in the post request.

Note: this was already working if the second parameter was sent as a
null or had problems.