feat(l2): add validium mode (#2365)
Validium is a [scaling
solution](https://ethereum.org/en/developers/docs/scaling/) that
enforces integrity of transactions using validity proofs like
[ZK-rollups](https://ethereum.org/en/developers/docs/scaling/zk-rollups/),
but doesn’t store transaction data on the Ethereum Mainnet.

**Description**
- Replace EIP 4844 transactions for EIP 1559
- Modify OnChainProposer contract so that it supports validium. It is
not the most efficient way of doing it but the simplest.
- Now the config.toml has a validium field.

Note: I'm not 100% sure about the changes that I made to the
OnChainProposer, there may be a mistake in the additions that I made. I
will review it though but I still consider worth opening this PR.


<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2313

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>