refactor(l1): implement code method for RLPxMessage enum and use it for encoding/decoding (#2454)
**Motivation**
Implements the refactor specified in the linked issues. Adding a single
`code` method for the RLPxMessage enum was not enough for both encoding
and decoding (as we would need to create the struct to call the method)
so an associated constant was also added to support both needs.
This solution fulfills the purpose of the issue, to have only one
instance of each message code that we can use for encoding/decoding of
messages, but its implementation is more complex than what we would have
liked. If the complexity is not acceptable, we should close both this PR
and the originating issue.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Implement `code` method for `RLPxMessage`
* Add `CODE` associated constant to `RLPxEncode` trait
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1035  (Also #1034)