chore(levm): minor fixes and refactors to previous PR (#2721)
**Motivation**

Fixing some comments left in a previous PR.

**Description**

Just opened this PR since some of the changes requested were out of
scope. Simply swapped some variables for constants in places where the
variables simply held the value of said constants, and applied a minor
change to an unwrap in the execution handlers.

Closes #2719