refactor(l2): replace sequencer config toml with CLI flags (#2606)
**Motivation**

- https://github.com/lambdaclass/ethrex/issues/2380
- https://github.com/lambdaclass/ethrex/issues/2574
- https://github.com/lambdaclass/ethrex/issues/2609

**Description**

- Adds CLI options for the sequencer components
- Extends `ethrex l2 init` options with `SequencerOptions` (a struct
that contains all the different components' options)
- Refactors `cmd/ethrex/l2.rs`
    - Moved the command code to `cmd/ethrex/l2/command.rs`.
    - Moved the command options to `cmd/ethrex/l2/options.rs`.
- Leaves the minimum necessary config in the
`sequencer_config_example.toml` (needed by the deployer).
- Leaves the minimum necessary logic in the
`crates/l2/utils/configs/toml_parser.rs` module (needed by the deployer
and prover).
- Adds CLI options for the contract deployer bin and the system
contracts updater bin (removing the need of a config file).
- Updates the L2 Makefile.
- Updates the Docker Compose files.
- Updates the `pr-main_l2` workflow.
- Updates the L2 integration test.
- Removes the `sequencer_config_example.toml` since it is not needed
anymore.
- Refactors the `crates/l2/contracts` module
- Renames the crate from `ethrex-l2_deployer` to `ethrex-l2_contracts`.
- Adds a `bin` module with the bins `ethrex_l2_l1_deployer` and
`ethrex_l2_system_contracts_updater`.
    - All the SDK-related logic was moved to the SDK lib.
- Cleans up the logic related to the config and toml parsing since now
the only bin relying on the config is the Prover. Everything relative to
the sequencer was removed, and now it is "hardcoded" for the Prover.

**How to test**

If you are in a dev environment, keep working as usual because under the
hood, the sequencer initialization is not relying anymore on the
`sequencer_config.toml`.

If you are in a prod environment, run `cargo run --release --features l2
-- l2 init --help` at the root of the repository to explore the
different configuration flags this PR adds.

**Caveats**

The prover config file is still needed by the prover (tracked in
https://github.com/lambdaclass/ethrex/issues/2576).

Closes #2574

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>