fix(l2): remove fork as parameter in get_state_transitions function (#2723)
In [this
PR](https://github.com/lambdaclass/ethrex/commit/a5da369f9a220976eb97429dcce0356363025d87)
we forgot to remove fork from some places in which get_state_transitions
was being used.

We should try to find a way to fix rust-analyzer so that it detects
those cases but I don't know how hard it is.