perf(l1): reduce transaction clone and Vec grow overhead in mempool (#2637)
**Motivation**

Improve perfomance

**Description**

Reduces transaction clone overhead on the mempool and Vec initial grow
overhead.

The main focus on this pr was the mempool fetch transaction method whose
overhead before was 15%~, reducing it to 13%~, whose main time is spent
in the filter transaction method, which had a clone taking 5% of the
time now reduced to non existent levels using an Arc.

Images of the perf profile:

Before


![image](https://github.com/user-attachments/assets/8a2a9e32-8e8d-4c24-8fd2-b5cb91401ee5)

After


![image](https://github.com/user-attachments/assets/1d3a0c56-5826-4efe-874b-8687c230c070)


<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number