fix(l1): support both `data` and `input` fields on `GenericTransaction` as long as they have the same value (#2685)
**Motivation**
Issue #2665 reported that some tools create transactions with both the
`data` and `input` fields set to the same value, which is not currently
supported by our deserialization which admits only one. Other
implementations also support having both fields in their equivalent of
`GenericTransaction`. This PR handles this case by adding a custom
deserialization that can parse both fields and check that they are equal
if set to prevent unexpected behaviours.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Implement custom field deserialization so that both `input` & `data`
fields are supported but deserialized as one
* Add test case for the reported failure case
* Use a non-trivial input for the current and added deserialization test
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2665