refactor(l2): replace prover config toml with CLI flags (#2771)
**Motivation**

We want to replace the .toml file used to configure the prover with a
cli

**Description**

- Remove all the code related to reading toml files
- Implement a struct ProverClientOptions that adds CLI options for the
prover

**How to test**

If you are in a dev environment, keep working as usual because under the
hood, the sequencer initialization is not relying anymore on the
prover_client_config.toml.

If you are in a prod environment, inside `crates/l2/prover` run `cargo
run --release -- --help` to explore the different configuration flags
this PR adds.

Closes #2576