refactor(l1): add `node_id` field to Node (node_id refactor 2/3) (#2786)
Based on #2778 
**Motivation**
Avoid constantly hashing the node's public key on kademlia operations by
adding `node_id` field. Before this PR we would hash the node's public
key every time we needed to add, remove or find a node in the kademlia
table, which is pretty often.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add `Node` field `node_id`
* Add `new` method for `Node` which handles node_id computation
* Use `node_id` for kademlia table (and some other) operations instead
of the public key so we no longer need to hash it when calculating the
bucket index (this affects most kademlia table reads/writes)

**Follow-Up Work**
Use `OnceLock` to cache for `node_id` computation (replacing the field
added by this PR) #2789
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number