refactor(l1): cache `node_id` computation (node_id refactor 3/3) (#2789)
Based on #2786 
**Motivation**
Use `OnceLock` to cache node_id computation so we only do it once but at
the same time don't need to do it unless we will use it. For example,
when we receive a Neighbours message we will decode all received nodes
but may not use them all if our kademlia table is full.
This PR can be ignored if we consider the cases where we would not need
to use a node's id scarce enough to not warrant the added complexity of
a cache. For example, the Neighbours case could be handled by using a
separate structure (without node_id) to decode the incoming node and
converting that to our Node (with node_id) when we insert that node into
our table.
The main consecuente of adding this cache is the `Node` no longer being
copy, which affects various areas of the networking codebase
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Convert public `Node` field `node_id: H256` into private field
`node_id: OnceLock<H256>`
* Add `Node` method `node_id`
* Fix code affected by `Node` no longer being `Copy`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number