feat(l1): properly calculate `enr` sequence field (#2679)
**Motivation**

The seq field in the node record was hardcoded with the unix time. 

<!-- Why does this pull request exist? What are its goals? -->

**Description**

The enr_seq field is updated by one when the node_record is changed. The
ping/pong messages are sent with the enr_seq in it, so the peer knows
when an update is made in the node_record. Since we don't modify the
node_record yet, the enr_seq is not being updated. There is a new PR
incoming (#2654) which is using this funtionality to inform the peers
about changes in the node_record.

A reference was added to the p2pcontext in order to be able to access
the current NodeRecord seq in several parts of the code.

Some functions firms were changed to accept this improvement.

A new config struct has been built to persist the enr seq field and also
store the known peers in the same file.

The test discv4::server::tests::discovery_enr_message checks this
feature

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

[enr](https://github.com/ethereum/devp2p/blob/master/enr.md)

Closes #1756