feat(l1): add eest hive tests to daily report (#2792)
**Motivation**

Have a better way to visualize the results from the execution of the EF
blockchain tests using Hive.

**Description**

Hive daily report now also runs the simulators
`ethereum/eest/consume-engine` and `ethereum/eest/consume-rlp` with the
blockchain fixtures of the `execution-spec-tests`. The version of the
fixtures is taken from `cmd/ef_tests/blockchain/.fixtures_url`.
This was also talked in #2474. 

Closes #2746 and part of #1988