feat(l1): capability negotation (#2840)
**Motivation**

Multiple version of the same protocol can be used when a connection is
established(eth/68 and eth/69 for example). At the moment, we can only
use one protocol version.

<!-- Why does this pull request exist? What are its goals? -->

**Description**

A vec of capability is used to pass multiple versions of the protocol to
some functions.

The struct RLPxConnection now stores capabilities struct instead of
numbers.

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->