refactor(levm): remove clones for account structs (#2684)
**Motivation**

Improving the performance of some cases by avoiding clones where
possible.

**Description**

Many clones of account structs were removed. This involved changing the
output of the get_account and access_account functions of the DB to
return a reference to an account, as well as refactorings of the code
which involved these functions.

Resolves [#2611](https://github.com/lambdaclass/ethrex/issues/2611)