fix(levm): fix eip 7702 logic around touched_accounts (#2859)
**Motivation**

- Fix error when executing a transaction of a block when syncing Holesky
in Prague by chaning behavior of the EVM.

**Description**

- We now set `code_address` and `bytecode` at the end of
`prepare_execution`. It's necessary because of EIP-7702.
- We change the place in which we add the delegated account to
`touched_accounts` → **CORE CHANGE**
- Change some outdated comments related to EIP7702 functions.
- Change `get_callee_and_code` to `get_callee` because we don't need the
code before `prepare_execution` and this is assigned afterwards.
- Create `set_code` function to CallFrame so that we calculate jump
destinations everytime we want to set the code, because it's always
necessary.


**In depth explanation: What was wrong with this transaction?**
The gas diff was 2000 between LEVM and REVM, but doing some math we
found out that the actual gas diff before refunds was 2500. The access
cost of accessing a COLD Address is 2600 and the cost of accessing a
WARM address is 100. 2600-100 = 2500. That's the difference between LEVM
and REVM, but where is it?
Reading EIP-7702 and analyzing our behavior made me realize:
(Capital Letters here are accounts)
- Transaction: A → B
- B had C as delegate account at the beginning of the transaction so we
add C to the `touched_accounts`.
- Transaction authority list sets B to have D as delegate, so that it's
not C anymore.
- During execution we make internal calls to C
- Our VM thinks C is in `touched_accounts` (that means warm) and
consumes 100 gas when accessing it instead of 2600.

Solution? Changing the moment in which we add the delegate account to
`touched_accounts`, so that we do it after the authorization list was
processed.