refactor(l1): add a generic VmDatabase trait to decouple vm from storage (#2801)
**Motivation**
I want to decouple the `vm` crate from the `storage` crate. This is a
first step in that direction.

**Description**
- Instead of depending on `Store` which is a type from storage, lets
crate a trait called VmDatabase that acts as an api for the vm to access
the state info.

Required follow ups:
- Move the StoreVmDatabase (previously StoreWrapper), which is the type
that implements `VmDatabase` from the `vm` crate to the blockchain
crate, which will be the connector between vm and storage.
- Move `ProverDatabase` to l2
- Move utility functions and types like `AccountUpdate` out of `storage`
and into `common`.
- Remove the dependency `ethrex_storage` from `ethrex_vm`.