fix(l1): added checks to newpayload and forkchoiceupdated (#2831)
**Motivation**

This should fix some sync and inconsistent behaviour problems with the
hive engine tests. The problem was happening when a sync process had
begun, and a block from that sync process entered the server. Some
checks for that scenario have been added.

Also made some tests in the CI easier to read and edit, while adding a
couple of them.

**Description**

- Made the CI tests have 1 by 1.
- Added some fixed CI tests to "Paris Engine tests" and "Engine
withdrawal tests"
- Added a check to forkchoiceupdatet for the body to be present before
forking.
- Added a check to execute_payload for the body of the parent to be
present before executing.

Fixes some tests in #1285