docs(levm): add docs explaining types of errors in the EVM (#2884)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Add docs about errors in the vm because maybe it isn't completely
clear which are propagated and which are not.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Note: I was thinking that maybe we should be more clear with our errors
struct. Maybe we should have a struct `LEVMError` that has inside 4
types of errors: `Internal`, `Database`, `TxValidation`, `EVM`. That way
it's easier to understand I think. (We should also do some clearing up
because it's quite messy and sometimes we don't even use the appropriate
errors I've seen)

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1537

I opened [this issue](https://github.com/lambdaclass/ethrex/issues/2886)
for refactoring errors because they are quite messy